handle admin user

+++ src/main/java/com/app/jdbc/smartrecruit/daos/UserDAO.java
@@ -79,9 +79,9 @@ public class UserDAO {
         }
     }
 
-    public int getUserCountByRole(Class<? extends User> role) {
+    public long getUserCountByRole(Class<? extends User> role) {
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
-            return session.createQuery("select count(u) from User u where type(u) = :role", Integer.class)
+            return session.createQuery("select count(u) from User u where type(u) = :role", Long.class)
                     .setParameter("role", role)
                     .uniqueResult();
         }
